vmbus_server: allow proxy driver to specify a device order (#2038)
This change allows vmbusproxy to specify a device order that will be
used to determine the offer order for offers provided by
`proxyintegration`. This means that in combination with the
`use_absolute_channel_order` option, vmbusproxy can control the order in
which devices get assigned channel IDs.

This is not a breaking change; although the
`VMBUS_PROXY_NEXT_ACTION_OUTPUT` structure has increased in size, older
versions of vmbusproxy.sys which don't support device order will simply
not write to the new field, which is therefore left at zero.

+++ vm/devices/vmbus/vmbus_channel/src/bus.rs
@@ -277,7 +277,7 @@ pub struct OfferParams {
     pub mnf_interrupt_latency: Option<Duration>,
     /// The order in which channels with the same interface will be offered to
     /// the guest (optional).
-    pub offer_order: Option<u32>,
+    pub offer_order: Option<u64>,
     /// Indicates whether the channel supports using encrypted memory for any
     /// external GPADLs and GPA direct ranges. This is only used when hardware
     /// isolation is in use.
